refactor(finyk): split utils.js into focused lib modules; move DB schema to migrations
Split 374-line utils.js into 8 single-responsibility modules under
src/modules/finyk/lib/: categories, formatting, accounts, debt,
transactions, goals, lsStats, spending. utils.js becomes a re-export
barrel — all existing import sites continue to work unchanged.

Move all inline CREATE TABLE statements from server/db.js ensureSchema()
into server/migrations/003_baseline_schema.sql (IF NOT EXISTS so safe
on existing DBs). ensureSchema() now only runs migrations; schema
history is fully tracked in schema_migrations.

642/642 tests pass, build green.
